requirements.txt
SQLAlchemy
psycopg2-binary
mypy
flask
pytest
alembic
ipdb
django
dj-database-url
ujson
bpython
celery
watchdog
requests
djangorestframework